+++ webpack.config.js
@@ -1,10 +1,8 @@
 const path = require('path');
-const webpack = require('webpack');
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-const LocalizationPlugin = require('./webpack/localizationPlugin');
 const parseParams = require('./webpack/parseParams.js');
-const { lstatSync, readdirSync, existsSync } = require('fs');
+const { lstatSync, readdirSync } = require('fs');
+const generateEntries = require('./webpack/generateEntries.js');
 
 const proxyPort = 8081;
 
@@ -17,46 +15,7 @@ module.exports = (env, argv) => {
     const getDirectories = source => readdirSync(source).map(name => path.join(source, name)).filter(isDirectory);
     const appsetupPaths = getDirectories(path.resolve(pathParam));
 
-    const entries = {};
-    const plugins = [
-        new webpack.IgnorePlugin(/^\.\/locale$/),
-        new CopyWebpackPlugin(
-            [
-                { from: 'resources', to: 'resources', context: __dirname },
-                { from: 'bundles/integration/admin-layerselector', to: 'assets/admin-layerselector', context: __dirname }
-            ]
-        )
-    ];
-
-    appsetupPaths.forEach(appDir => {
-        const minifierPath = path.resolve(appDir, 'minifierAppSetup.json');
-        const mainJsPath = path.resolve(appDir, 'main.js');
-        let targetPath;
-        if (existsSync(minifierPath)) {
-            targetPath = minifierPath;
-        } else if (existsSync(mainJsPath)) {
-            targetPath = mainJsPath;
-        } else {
-            // skip
-            console.log('No minifierAppSetup.json file or main.js file in ' + appDir + '. Skipping!');
-            return;
-        }
-        const appName = path.basename(appDir);
-        const copyPlugin = new CopyWebpackPlugin(
-            [
-                { from: appDir, to: appName },
-                { from: 'resources/icons.css', to: appName, context: __dirname },
-                { from: 'resources/icons.png', to: appName, context: __dirname }
-            ]
-        );
-        entries[appName] = [
-            path.resolve(__dirname, './webpack/polyfill.js'),
-            path.resolve(__dirname, './webpack/oskari-core.js'),
-            targetPath
-        ];
-        plugins.push(copyPlugin);
-        plugins.push(new LocalizationPlugin(appName));
-    });
+    const {entries, plugins} = generateEntries(appsetupPaths, __dirname);
 
     // Common config for both prod & dev
     const config = {
